+++ spec/deliveries/admin_delivery_spec.rb
@@ -3,14 +3,14 @@
 describe AdminDelivery, type: :delivery do
   let(:job) { double }
 
-  describe '#job_execution_report' do
-    before do
-      allow(Pullmetry::Application).to(
-        receive(:credentials).and_return({ reports_webhook_url: 'https://hooks.slack.com/services/T0/B0/G0' })
-      )
-    end
+  before do
+    allow(Rails.application.credentials).to(
+      receive(:[]).with(:reports_webhook_url).and_return('https://hooks.slack.com/services/T0/B0/G0')
+    )
+  end
 
-    it 'delivers to slack_webhook', skip: 'GA does not recognize credentials' do
+  describe '#job_execution_report' do
+    it 'delivers to slack_webhook' do
       expect {
         described_class.with(job_name: job.class.name).job_execution_report.deliver_later
       }.to deliver_via(:slack_webhook)
@@ -20,13 +20,7 @@ describe AdminDelivery, type: :delivery do
   describe '#feedback_created' do
     let!(:feedback) { create :feedback }
 
-    before do
-      allow(Pullmetry::Application).to(
-        receive(:credentials).and_return({ reports_webhook_url: 'https://hooks.slack.com/services/T0/B0/G0' })
-      )
-    end
-
-    it 'delivers to slack_webhook', skip: 'GA does not recognize credentials' do
+    it 'delivers to slack_webhook' do
       expect {
         described_class.with(id: feedback.id).feedback_created.deliver_later
       }.to deliver_via(:slack_webhook)